feat: terrain contact class registration in class solver and including the interpolation density level set functions in class interpolation
