+++ inc/Solver/Solver.h
@@ -14,6 +14,7 @@ using std::vector;
 class Body;
 class Mesh;
 class Particle;
+class TerrainContact;
 
 /// \class Solver
 /// \brief Represents the operations to solve the equations in time.
@@ -44,19 +45,18 @@ public:
 	/// \brief Register the particles in the solver
 	inline void registerParticles(vector<Particle*>* p){ this->particles=p; }
 
+	/// \brief Register the terrain contact in the solver
+	inline void registerTerrainContact(TerrainContact* terrainContact){ this->terrainContact=terrainContact; }
+
 protected:
 
 	Mesh* mesh; //!< pointer to mesh
 	vector<Body*>* bodies; //!< pointer to bodies
 	vector<Particle*>* particles; //!< pointer to particles
+	TerrainContact* terrainContact; //!< pointer to terrain contact
 };
 
-inline Solver::Solver():mesh(0),bodies(0) {
-	
-}
-
-inline Solver::~Solver() {
-	
-}
+inline Solver::Solver():mesh(0),bodies(0),particles(0),terrainContact(0) { }
+inline Solver::~Solver() { }
 
 #endif /* SOLVER_H_ */
